requirements.txt
flask
Flask-SQLAlchemy
flask-login
flask-migrate
matplotlib
beautifulsoup4
requests
notebook
